pyproject.toml
[project]
name = "ai-vs-insan"
version = "0.1.0"
description = "AI vs Insan: Ocak Yemini — Voice AI Hackathon 2026"
requires-python = ">=3.13"
dependencies = [
    "fal-client>=0.13.0",
    "httpx>=0.28.1",
    "langchain-core>=1.2.9",
    "langgraph>=1.0.8",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
]
